feature(cli): complete CLI pivot with argc/argv dispatch - Rewrote main.c as minimal CLI entry point - Implemented parse_and_dispatch with function pointer dispatch - Added static helpers: parrse_double, dispatch_unary, dispatch_binary - Added mathinator_exit_code enum with 0/1/2 exit codes - Full result_t pipeline with MATHINATOR_OK/INVALID/UNDEF - Commands: add, subtract, multiply, divide, remainder, power, log, sqroot, factorial - stderr for errors, stdout for results - Handles domain errors, overflow, invalid input, unknown commands

+++ include/math/arithmetic.h
@@ -5,10 +5,10 @@ double add(double x, double y);      // Adds two numbers and returns the result
 double subtract(double x, double y); // Subtracts y from x and returns the result
 double product(double x, double y);  // Multiplies two numbers and returns the result
 double divide(double x, double y);   // Divides x by y and returns the result
-int rem(int x, int y);             // Calculates the remainder of x divided by y
+double rem(double x, double y);             // Calculates the remainder of x divided by y
 double sqroot(double x);            // Computes the square root of x
 double power(double x, double y);    // Raises x to the power of y
-double fact(int x);                   // Calculates the factorial of x
+double fact(double x);                   // Calculates the factorial of x
 double logarithm(double x, double b);   // Log of (x) to the base b
 
 #endif
